Move fusions out of subfolder
